[refactor] #22 후보자 목록 코드 조회 client 패키지로 분리
